remove utility file to co-location register page
